Created endpoints for Product/ProductType

+++ src/main/java/com/example/miracles_store/entity/Product.java
@@ -12,6 +12,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
@@ -28,7 +30,7 @@ public class Product {
 
     private String description;
 
-    private double cost;
+    private BigDecimal cost;
 
     @ManyToOne
     @JoinColumn(name = "product_type_id")
